+++ lib/src/existing_steps.dart
@@ -1,9 +1,9 @@
 import 'dart:io';
 
-import 'package:bdd_widget_test/src/constants.dart';
 import 'package:path/path.dart' as p;
 
-Map<String, String> getExistingStepSubfolders(String featureDir) {
+Map<String, String> getExistingStepSubfolders(
+    String featureDir, String stepFolderName) {
   final stepFolder = p.join(featureDir, stepFolderName);
   final steps = Directory(stepFolder);
   if (!steps.existsSync()) {
@@ -12,12 +12,15 @@ Map<String, String> getExistingStepSubfolders(String featureDir) {
   return steps.listSync(recursive: true).asMap().map(
         (key, value) => MapEntry<String, String>(
           p.basename(value.path),
-          _getStepFolder(value.uri.pathSegments),
+          _getStepFolder(
+            value.uri.pathSegments,
+            stepFolderName,
+          ),
         ),
       );
 }
 
-String _getStepFolder(List<String> pathSegments) {
+String _getStepFolder(List<String> pathSegments, String stepFolderName) {
   final stepFolderIndex = pathSegments.indexOf(stepFolderName);
   final stepFolder =
       pathSegments.getRange(stepFolderIndex, pathSegments.length - 1);