+++ lib/src/generator_options.dart
@@ -1,18 +1,23 @@
-const defaultTestName = 'testWidgets';
+const _defaultTestName = 'testWidgets';
+const _stepFolderName = 'step';
 
 class GeneratorOptions {
   const GeneratorOptions({
     String? testMethodName,
     List<String>? externalSteps,
-  })  : testMethodName = testMethodName ?? defaultTestName,
+    String? stepFolderName,
+  })  : stepFolder = stepFolderName ?? _stepFolderName,
+        testMethodName = testMethodName ?? _defaultTestName,
         externalSteps = externalSteps ?? const [];
 
   factory GeneratorOptions.fromMap(Map<String, dynamic> json) =>
       GeneratorOptions(
         testMethodName: json['testMethodName'] as String?,
         externalSteps: (json['externalSteps'] as List?)?.cast<String>(),
+        stepFolderName: json['stepFolderName'] as String?,
       );
 
+  final String stepFolder;
   final String testMethodName;
   final List<String> externalSteps;
 }